Organize typings and fix testings
