+++ package.json
@@ -36,7 +36,6 @@
     "bcrypt-nodejs": "^0.0.3",
     "cors": "^2.8.5",
     "form-data": "^3.0.0",
-    "graphql": "^14.6.0",
     "graphql-import": "^0.7.1",
     "graphql-middleware": "^4.0.2",
     "graphql-tools": "^5.0.0",
@@ -92,8 +91,5 @@
     "ts-jest": "^25.2.1",
     "ts-node": "^8.6.2",
     "typescript": "3.8.3"
-  },
-  "resolutions": {
-    "graphql": "14.6.0"
   }
 }
