+++ src/types.ts
@@ -1,0 +1,14 @@
+export interface JwtUser {
+  userId: string;
+  role: number;
+  iat: number;
+}
+export enum Role {
+  User,
+  Admin,
+}
+
+export interface Auth {
+  id: string;
+  role: Role;
+}
